LayoutTests/fast/events/wheel/wheel-event-listeners-on-document-made-passive-expected.txt

Tests that a wheel event handler on the document is passive (cannot prevent scrolling)
PASS windowScrollEventCount > 0 is true
PASS defaultPrevented is false
PASS successfullyParsed is true

TEST COMPLETE
